improved user experience
Now the script will check if there are already AppDir related to the package you want to create. If it exists, the creation of the AppImage will resume where it left off.
The creation of AppImage can be repeated in case of malfunctions of the previously created AppImage: added a question (Y / N), if Y you will be offered the possibility to export the AppDir again in AppImage, otherwise press N to close the program.